Make sonar happy

+++ app/src/main/java/me/devsaki/hentoid/fragments/queue/QueueFragment.java
@@ -496,7 +496,7 @@ public class QueueFragment extends Fragment implements ItemTouchCallback, Simple
             dlPreparationProgressBar.setVisibility(View.GONE);
         }
 
-        dlPreparationProgressBar.setProgress1(event.total - event.done);
+        dlPreparationProgressBar.setProgress1((float) (event.total - event.done));
     }
 
     /**
